# nodejs-server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Estuary API
  description: This is the API for the Estuary application.
  termsOfService: http://estuary.tech
  contact:
    name: API Support
    url: https://docs.estuary.tech/feedback
  license:
    name: Apache 2.0 Apache-2.0 OR MIT
    url: https://github.com/application-research/estuary/blob/master/LICENSE.md
  version: 0.0.0
servers:
- url: //api.estuary.tech/
security:
- bearerAuth: []
paths:
...
  /deal/estimate:
    post:
      tags:
      - deals
      summary: Estimate the cost of a deal
      description: This endpoint estimates the cost of a deal
      operationId: dealEstimatePOST
      requestBody:
        description: "The size of the deal in bytes, the replication factor, and the\
          \ duration of the deal in blocks"
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/main.estimateDealBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-codegen-request-body-name: body
      x-swagger-router-controller: Deals
  /deal/info/{dealid}:
    get:
      tags:
      - deals
      summary: Get Deal Info
      description: This endpoint returns the deal info for a deal
      operationId: dealInfoDealidGET
      parameters:
      - name: dealid
        in: path
        description: Deal ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /deal/proposal/{propcid}:
    get:
      tags:
      - deals
      summary: Get Proposal
      description: This endpoint returns the proposal for a deal
      operationId: dealProposalPropcidGET
      parameters:
      - name: propcid
        in: path
        description: Proposal CID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /deal/query/{miner}:
    get:
      tags:
      - deals
      summary: Query Ask
      description: This endpoint returns the ask for a given CID
      operationId: dealQueryMinerGET
      parameters:
      - name: miner
        in: path
        description: CID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /deal/status-by-proposal/{propcid}:
    get:
      tags:
      - deals
      summary: Get Deal Status by PropCid
      description: Get Deal Status by PropCid
      operationId: dealStatus_by_proposalPropcidGET
      parameters:
      - name: propcid
        in: path
        description: PropCid
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /deal/status/{miner}/{propcid}:
    get:
      tags:
      - deals
      summary: Deal Status
      description: This endpoint returns the status of a deal
      operationId: dealStatusMinerPropcidGET
      parameters:
      - name: miner
        in: path
        description: Miner
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: propcid
        in: path
        description: Proposal CID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /deal/transfer/in-progress:
    get:
      tags:
      - deals
      summary: Transfer In Progress
      description: This endpoint returns the in-progress transfers
      operationId: dealTransferIn_progressGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /deal/transfer/status:
    post:
      tags:
      - deals
      summary: Transfer Status
      description: This endpoint returns the status of a transfer
      operationId: dealTransferStatusPOST
      requestBody:
        description: Channel ID
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/main.ChannelIDParam'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-codegen-request-body-name: chanid
      x-swagger-router-controller: Deals
...
  /net/addrs:
    get:
      tags:
      - net
      summary: Net Addrs
      description: This endpoint is used to get net addrs
      operationId: netAddrsGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Net
...
components:
  schemas:
...
    main.ChannelIDParam:
      type: object
      properties:
        id:
          type: integer
        initiator:
          type: string
        responder:
          type: string
...
    main.estimateDealBody:
      type: object
      properties:
        durationBlks:
          type: integer
        replication:
          type: integer
        size:
          type: integer
        verified:
          type: boolean
...
    util.HttpError:
      type: object
      properties:
        code:
          type: integer
        details:
          type: string
        reason:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header
...